Add files via upload
Similar to last projectile motion simulation, introduction of acceleration to the projectile made it slightly more complex and interesting to play around with parameters. Unlike constant velocity simulation, which allows for nice parabolic shape, achieving the same shape with changing velocity is much harder. Value for acceleration for projectile to hit the ground after 10 seconds ~ 6.94 m/s^2 which is interestingly same value as for the velocity in last example. Maybe my maths or code was wrong, or both.