Add files via upload
Here I used my knowledge of ODE's to solve an equation with complex roots that describes the motion of a hanging mass subject to dampening, spring constant and an applied external force. I struggle to tell from the graph but I understand we expect to see the object eventually oscillate at the frequency of the particular solution as t goes to infinity. Overall was not to hard to simulate, although I did need some help trying to plot the function, was not as straight forward as I expected.